Mars Lander (optimisation) solution 1.0
The solution to the "Mars Lander" Codingame puzzle that uses quadratic bezier curves to guide the craft to the surface. This solution gives 2080 remaining fuel (which is 30th UK and 249th Worldwide at the time of writing). Reach out to me if you have suggestions/ spot problems that can improve this code.

Anay